Remove DatabaseIntegration sample. Merge IastController
