--- tracer/test/test-applications/security/Samples.Security.AspNetCore5.DatabaseIntegration/Data/ApplicationDbContext.cs
@@ -1,15 +1,0 @@
-﻿#if NET7_0_OR_GREATER
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
-
-namespace Samples.Security.AspNetCore5.Data.DatabaseIntegration
-{
-    public class ApplicationDbContext : IdentityDbContext
-    {
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
-            : base(options)
-        {
-        }
-    }
-}
-#endif
