--- tracer/test/test-applications/security/Samples.Security.AspNetCore5.DatabaseIntegration/Properties/launchSettings.json
@@ -1,131 +1,0 @@
-{
-  "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DD_APPSEC_ENABLED": "true",
-        "DD_EXPERIMENTAL_APPSEC_STANDALONE_ENABLED": "true",
-        "DD_APPSEC_BLOCKING_ENABLED": "false",
-        "DD_DOTNET_TRACER_HOME": "c:\\code\\dd-trace-dotnet\\shared\\bin\\monitoring-home",
-        "CORECLR_ENABLE_PROFILING": "1",
-        "CORECLR_PROFILER": "{846F5F1C-F9AE-4B07-969E-05C26BC060D8}",
-        "CORECLR_PROFILER_PATH_32": "c:\\code\\dd-trace-dotnet\\shared\\bin\\monitoring-home\\win-x86\\Datadog.Tracer.Native.dll",
-        "CORECLR_PROFILER_PATH_64": "c:\\code\\dd-trace-dotnet\\shared\\bin\\monitoring-home\\win-x64\\Datadog.Tracer.Native.dll",
-        "DD_VERSION": "1.0.0",
-        "DD_TRACE_HEADER_TAGS": "sample.correlation.identifier, Server",
-        "DD_IAST_ENABLED": "true",
-        "DD_IAST_DEDUPLICATION_ENABLED": "false",
-        "DD_TRACE_DEBUG": "1"
-      },
-      "nativeDebugging": true,
-      "use64Bit": true
-    },
-    "Samples.Security.AspNetCore5": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DD_APPSEC_ENABLED": "true",
-        "DD_EXPERIMENTAL_APPSEC_STANDALONE_ENABLED": "true",
-        "DD_APPSEC_BLOCKING_ENABLED": "true",
-        "DD_APPSEC_RASP_ENABLED": "true",
-        "DD_DOTNET_TRACER_HOME": "c:\\code\\dd-trace-dotnet\\\\shared\\bin\\monitoring-home\\",
-        "CORECLR_ENABLE_PROFILING": "1",
-        "CORECLR_PROFILER": "{846F5F1C-F9AE-4B07-969E-05C26BC060D8}",
-        "CORECLR_PROFILER_PATH_32": "c:\\code\\dd-trace-dotnet\\\\shared\\bin\\monitoring-home\\win-x86\\Datadog.Tracer.Native.dll",
-        "CORECLR_PROFILER_PATH_64": "c:\\code\\dd-trace-dotnet\\\\shared\\bin\\monitoring-home\\win-x64\\Datadog.Tracer.Native.dll",
-        "CORECLR_PROFILER_PATH_ARM64": "c:\\code\\dd-trace-dotnet\\\\shared\\bin\\monitoring-home\\win-arm64\\Datadog.Tracer.Native.dll",
-        "DD_VERSION": "1.0.0",
-        "DD_TRACE_HEADER_TAGS": "sample.correlation.identifier, Server",
-        "DD_TRACE_DEBUG": "1",
-        "DD_SERVICE": "Samples.Security.Netcore5",
-        "DD_ENV": "asm-samples",
-        "DD_IAST_ENABLED": "true",
-        "DD_IAST_DEDUPLICATION_ENABLED": "false",
-        "DD_IAST_VULNERABILITIES_PER_REQUEST": "100",
-        "DD_IAST_REQUEST_SAMPLING": "100",
-        "DD_IAST_MAX_CONCURRENT_REQUESTS": "1",
-        "DD_WRITE_INSTRUMENTATION_TO_DISK": "false",
-        "DD_API_SECURITY_ENABLED": "true",
-        "DD_API_SECURITY_MAX_CONCURRENT_REQUESTS": "100"
-      },
-      "Logging": {
-        "LogLevel": {
-          "Default": "Warning",
-          "Microsoft": "Warning",
-          "Microsoft.Hosting.Lifetime": "Warning"
-        }
-      },
-      "nativeDebugging": true,
-      "use64Bit": false,
-      "applicationUrl": "http://localhost:5458"
-    },
-    "Linux": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DD_DOTNET_TRACER_HOME": "$(SolutionDir)tracer/bin/tracer-home",
-        "CORECLR_ENABLE_PROFILING": "1",
-        "CORECLR_PROFILER": "{846F5F1C-F9AE-4B07-969E-05C26BC060D8}",
-        "CORECLR_PROFILER_PATH": "$(SolutionDir)shared/bin/monitoring-home/tracerDatadog.Trace.ClrProfiler.Native.so",
-        "DD_VERSION": "1.0.0",
-        "DD_TRACE_HEADER_TAGS": "sample.correlation.identifier, Server",
-        "DD_TRACE_DEBUG": "0"
-      },
-      "Logging": {
-        "LogLevel": {
-          "Default": "Info",
-          "Microsoft": "Info",
-          "Microsoft.Hosting.Lifetime": "Info"
-        }
-      },
-      "nativeDebugging": true,
-      "use64Bit": true,
-      "applicationUrl": "http://localhost:5458"
-    },
-    "OSX": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DD_DOTNET_TRACER_HOME": "$(SolutionDir)shared/bin/monitoring-home",
-        "CORECLR_ENABLE_PROFILING": "1",
-        "CORECLR_PROFILER": "{846F5F1C-F9AE-4B07-969E-05C26BC060D8}",
-        "CORECLR_PROFILER_PATH": "$(SolutionDir)shared/bin/monitoring-home/osx/Datadog.Trace.ClrProfiler.Native.dylib",
-        "DD_VERSION": "1.0.0",
-        "DD_TRACE_HEADER_TAGS": "sample.correlation.identifier, Server",
-        "DD_TRACE_DEBUG": "1",
-        "DD_SERVICE" : "Samples.Security.AspNetCore5.OSX",
-        "DD_ENV": "asm-samples",
-        "DD_APPSEC_ENABLED": "true",
-        "DD_APPSEC_BLOCKING_ENABLED": "true",
-        "DD_IAST_ENABLED": "true",
-        "DD_IAST_DEDUPLICATION_ENABLED": "false",
-        "DD_IAST_VULNERABILITIES_PER_REQUEST": "100",
-        "DD_IAST_REQUEST_SAMPLING": "100",
-        "DD_IAST_MAX_CONCURRENT_REQUESTS": "1"
-      },
-      "Logging": {
-        "LogLevel": {
-          "Default": "Warning",
-          "Microsoft": "Warning",
-          "Microsoft.Hosting.Lifetime": "Warning"
-        }
-      },
-      "nativeDebugging": true,
-      "use64Bit": true,
-      "applicationUrl": "http://localhost:5458"
-    }
-  },
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:5458/",
-      "sslPort": 44320
-    }
-  }
-}
