--- tracer/test/test-applications/security/Samples.Security.AspNetCore5.DatabaseIntegration/Views/Iast/Email.cshtml.cs
@@ -1,12 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace Samples.Security.AspNetCore5.Views.Iast
-{
-    public class EmailModel : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
