--- tracer/test/test-applications/security/Samples.Security.AspNetCore5.DatabaseIntegration/Views/_ViewImports.cshtml
@@ -1,2 +1,0 @@
-﻿@using Samples.Security.AspNetCore5
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
