--- tracer/test/test-applications/security/Samples.Security.AspNetCore5.DatabaseIntegration/Views/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "_Layout";
-}
