--- tracer/test/test-applications/security/Samples.Security.AspNetCore5.DatabaseIntegration/appsettings.json
@@ -1,7 +1,0 @@
-{
-  "ConnectionStrings": {
-    "DefaultConnection": "Data Source=Data\\app.db"
-  },
-  "AllowedHosts": "*",
-  "CreateDb": false
-}
