--- tracer/test/test-applications/security/Samples.Security.AspNetCore5.DatabaseIntegration/web.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <location path="." inheritInChildApplications="false">
-    <system.webServer>
-      <handlers>
-        <add name="aspNetCore" path="*" verb="*" modules="AspNetCoreModuleV2" resourceType="Unspecified" />
-      </handlers>
-
-      <!-- Set in applicationHost.config instead -->
-      <!-- <aspNetCore processPath="%LAUNCHER_PATH%" arguments="%LAUNCHER_ARGS%" stdoutLogEnabled="false" stdoutLogFile=".\stdout" hostingModel="%LAUNCHER_HOSTING_MODEL%" /> -->
-    </system.webServer>
-  </location>
-</configuration>
