+++ tracer/test/test-applications/security/Samples.Security.AspNetCore5/Data/IastControllerHelper.cs
@@ -1,14 +1,47 @@
 using System;
+using System.Data.SqlClient;
 using System.Data.SQLite;
+using System.Linq;
 using System.Text;
 #if NETCOREAPP
 using Microsoft.Data.Sqlite;
+#endif
+#if NET5_0_OR_GREATER
+using MySql.Data.MySqlClient;
+using Npgsql;
+using Oracle.ManagedDataAccess.Client;
 #endif
 
 namespace Samples.Security.Data
 {
     public static class IastControllerHelper
     {
+        private static readonly string[] SqlServerTables =
+            [
+                "CREATE TABLE Persons (Id uniqueidentifier NOT NULL,Name varchar(50) NOT NULL,Surname varchar(50) NULL,Gender char(10) NULL,Email varchar(50) NULL,FlattedAddress varchar(500) NULL,Mobile varchar(50) NULL,NIF varchar(50) NULL,Phone varchar(50) NULL,PostalCode varchar(50) NULL,Details varchar(5000) NULL,CityIniqueId varchar(50) NULL,CountryUniqueId varchar(50) NULL,ImagePath varchar(500) NULL)",
+                @"CREATE TABLE Books(Id uniqueidentifier NOT NULL,Title nvarchar(500) NOT NULL,Price numeric(18, 2) NULL,ISBN nvarchar(50) NULL,	Pages int NOT NULL,    Author nvarchar(500) NOT NULL,    Editorial nvarchar(50) NULL,	Prologue nvarchar(4000) NULL,	Format int NULL)",
+            ];
+
+        private static readonly string[] PostgresTables =
+            [
+                "DROP TABLE IF EXISTS Persons, Books;",
+                "CREATE TABLE Persons (Id uuid NOT NULL,Name varchar(50) NOT NULL,Surname varchar(50) NULL,Gender char(10) NULL,Email varchar(50) NULL,FlattedAddress varchar(500) NULL,Mobile varchar(50) NULL,NIF varchar(50) NULL,Phone varchar(50) NULL,PostalCode varchar(50) NULL,Details varchar(5000) NULL,CityIniqueId varchar(50) NULL,CountryUniqueId varchar(50) NULL,ImagePath varchar(500) NULL)",
+                @"CREATE TABLE Books(Id uuid NOT NULL,Title varchar(500) NOT NULL,Price numeric(18, 2) NULL,ISBN varchar(50) NULL,	Pages int NOT NULL,    Author varchar(500) NOT NULL,    Editorial varchar(50) NULL,	Prologue varchar(4000) NULL,	Format int NULL)",
+            ];
+
+        private static readonly string[] MySqlTables =
+        [
+            "DROP TABLE IF EXISTS Persons, Books;",
+            "CREATE TABLE Persons (Id char(36) NOT NULL,Name varchar(50) NOT NULL,Surname varchar(50) NULL,Gender char(10) NULL,Email varchar(50) NULL,FlattedAddress varchar(500) NULL,Mobile varchar(50) NULL,NIF varchar(50) NULL,Phone varchar(50) NULL,PostalCode varchar(50) NULL,Details varchar(5000) NULL,CityIniqueId varchar(50) NULL,CountryUniqueId varchar(50) NULL,ImagePath varchar(500) NULL)",
+            @"CREATE TABLE Books(Id char(36) NOT NULL,Title varchar(500) NOT NULL,Price numeric(18, 2) NULL,ISBN varchar(50) NULL,	Pages int NOT NULL,    Author varchar(500) NOT NULL,    Editorial varchar(50) NULL,	Prologue varchar(4000) NULL,	Format int NULL)",
+        ];
+
+        private static readonly string[] OracleTables =
+        [
+            "DROP TABLE IF EXISTS Persons",
+            "CREATE TABLE Persons (Id char(40) NOT NULL,Name varchar(50) NOT NULL,Surname varchar(50) NULL,Gender char(10) NULL,Email varchar(50) NULL,FlattedAddress varchar(500) NULL,Mobile varchar(50) NULL,NIF varchar(50) NULL,Phone varchar(50) NULL,PostalCode varchar(50) NULL,Details varchar(2000) NULL,CityIniqueId varchar(50) NULL,CountryUniqueId varchar(50) NULL,ImagePath varchar(500) NULL)",
+        ];
+
         private static readonly string[] SqliteTables =
         [
             "CREATE TABLE Persons (Id NOT NULL,Name varchar(50) NOT NULL,Surname varchar(50) NULL,Gender char(10) NULL,Email varchar(50) NULL,FlattedAddress varchar(500) NULL,Mobile varchar(50) NULL,NIF varchar(50) NULL,Phone varchar(50) NULL,PostalCode varchar(50) NULL,Details varchar(5000) NULL,CityIniqueId varchar(50) NULL,CountryUniqueId varchar(50) NULL,ImagePath varchar(500) NULL)",
@@ -25,6 +58,118 @@ namespace Samples.Security.Data
             "INSERT INTO Books (Id, Title, Price, ISBN, Pages, Author, Editorial, Prologue, Format) VALUES ('026decfd-bba3-4aa8-85d4-1c71ffcfe8e9', 'CLR via C#', CAST(50.00 AS Decimal(18, 2)), '0735669954', 894, ' Jeffrey Richter', 'Microsoft Press', 'bla bla bla', 0)"
         ];
 
+#if NET5_0_OR_GREATER
+        public static OracleConnection CreateOracleDatabase()
+        {
+            var host = Environment.GetEnvironmentVariable("ORACLE_HOST") ?? "localhost";
+            var cstringBuilder = new OracleConnectionStringBuilder();
+            cstringBuilder.DataSource = $"{host}:1521/FREE";
+            cstringBuilder.UserID = "system";
+            cstringBuilder.Password = "testpassword";
+
+            var connection = new OracleConnection(cstringBuilder.ConnectionString);
+            connection.Open();
+
+            foreach (var command in OracleTables)
+            {
+                new OracleCommand(command, connection).ExecuteNonQuery();
+            }
+            foreach (var command in PopulateCommands.Take(5))
+            {
+                new OracleCommand(command, connection).ExecuteNonQuery();
+            }
+
+            return connection;
+        }
+
+        public static MySqlConnection CreateMySqlDatabase()
+        {
+            var connectionString = Environment.GetEnvironmentVariable("MYSQL_CONNECTION_STRING");
+
+            if (connectionString == null)
+            {
+                var host = Environment.GetEnvironmentVariable("MYSQL_HOST") ?? "localhost";
+                var port = Environment.GetEnvironmentVariable("MYSQL_PORT") ?? "3307";
+                connectionString = $"server={host};user=mysqldb;password=mysqldb;port={port};database=world";
+            }
+
+            var connection = new MySqlConnection(connectionString);
+            connection.Open();
+
+            foreach (var command in MySqlTables)
+            {
+                new MySqlCommand(command, connection).ExecuteNonQuery();
+            }
+            foreach (var command in PopulateCommands)
+            {
+                new MySqlCommand(command, connection).ExecuteNonQuery();
+            }
+
+            return connection;
+        }
+
+        public static NpgsqlConnection CreatePostgresDatabase()
+        {
+            // assumes the postgres docker image is running...
+            var connectionString = Environment.GetEnvironmentVariable("POSTGRES_CONNECTION_STRING");
+
+            if (connectionString == null)
+            {
+                var host = Environment.GetEnvironmentVariable("POSTGRES_HOST") ?? "localhost";
+                connectionString = $"Host={host};Username=postgres;Password=postgres;Database=postgres";
+            }
+            var connection = new NpgsqlConnection(connectionString);
+            connection.Open();
+
+            foreach (var command in PostgresTables)
+            {
+                new NpgsqlCommand(command, connection).ExecuteNonQuery();
+            }
+            foreach (var command in PopulateCommands)
+            {
+                new NpgsqlCommand(command, connection).ExecuteNonQuery();
+            }
+
+            return connection;
+        }
+
+        public static SqlConnection CreateSqlServerDatabase()
+        {
+            // assumes the sqlserver docker image is running...
+            var connectionString = Environment.GetEnvironmentVariable("SQLSERVER_CONNECTION_STRING") ??
+                                   @"Server=localhost;User Id=sa;Password=Strong!Passw0rd;";
+
+            int numAttempts = 3;
+            for (int i = 0; i < numAttempts; i++)
+            {
+                SqlConnection connection = null;
+                try
+                {
+                    connection =new SqlConnection(connectionString);
+                    connection.Open();
+
+                    foreach (var command in SqlServerTables)
+                    {
+                        new SqlCommand(command, connection).ExecuteNonQuery();
+                    }
+                    foreach (var command in PopulateCommands)
+                    {
+                        new SqlCommand(command, connection).ExecuteNonQuery();
+                    }
+                    return connection;
+                }
+                catch (Exception e)
+                {
+                    Console.WriteLine(e);
+                    connection?.Dispose();
+                }
+
+            }
+
+            throw new Exception($"Unable to open connection to connection string {connectionString} after {numAttempts} attempts");
+        }
+
+#endif
         public static SQLiteConnection CreateSystemDataDatabase()
         {
             var builderSystemData = new SQLiteConnectionStringBuilder { DataSource = ":memory:" };
